requirements.txt
Babel==2.4.0
certifi==2017.7.27.1
chardet==3.0.4
Django==1.11
django-phonenumber-field==1.3.0
docopt==0.6.1
docopts==0.6.1
geopy==1.11.0
idna==2.6
phonenumberslite==8.7.1
psycopg2==2.7.3
python-decouple==3.0
pytz==2017.2
requests==2.18.4
requests-toolbelt==0.8.0
rets-python==0.3.9
titlecase==0.11.0
udatetime==0.0.13
urllib3==1.22
